requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
apscheduler==3.10.4
aiosqlite==0.20.0
